Improve lookup rules.
This rather large patch implements new lookup rules which attempt to mirror the
behaviour defined in Ada (at least for the constructs which map over to Comma).

In addition, the ParseClient API has changed where function and procedure calls
take a Node for their connective (instead of an IdentifierInfo).   Named
subroutine calls have their connectives processed with acceptFunctionName or
acceptProcedureName, with the result being passed to the new acceptFunctionCall
and acceptProcedureCall methods.  This change paves the way for more complex
connectives in the future (e.g. first class functions).

In addition, a small bug was fixed in Parser::seekEndIf causing an infinite loop
when trying to walk out of an if statement.