Introduce virtual Type::getDeclaration method.
Previously, many of the type nodes provided a getDeclaration method with a
return type suited to implementing class.  However, a virtual method in Type is
quite useful.  We cannot use covariant return types here since the declaration
hierarchy is not in place (forward declarations only) when Type.h is included.