Initial support for exceptions.
This gives Comma initial runtime support for DWARF zero cost exceptions.  This
is certainly WIP.  There is no support for user defined exceptions or handlers
yet.  Rather, the main stub is codegened with a catch-all to manage system
generated exceptions.

To test the system, support to emit simple range checks of scalar assignments is
in place.