Parse prj and inj as names using a function call like syntax.
Previously, prj and inj were low-precedence prefix operators.  For example,
given a carrier type of Integer one would write:

    function Foo (S : Integer) return % is
    begin
       return prj S + 1;
    end Foo;

The problem with that syntax is that there was no way to apply an attribute to
the result of prj or inj since they did not denote names in the grammar.  Also,
needing to write (prj X) to control the precedence was awkward.

The return statement in the example above now reads as:

    return prj(N + 1);

and since the application of prj or inj are parsed as names, attributes may be
applied to the result of the view conversion.