Move choice ordering predicates from AggregateChecker to KeyedAggExpr.
Codegen needs to perform sorting on keyed aggregates too.  Unfortunately, a full
API in KeyedAggExpr for this task will not work for typecheck since the type of
the aggregate is not known to the AST at the point the sorting needs to be done
(the type is needed to decide between signed and unsigned comparisons).  This
solution is not ideal but is far better than duplicating code.