Extend support for exceptions.
Define the ExceptionDecl class.  Currently unused, but will serve to represent
language and user defined exception entities.

Generalize libruntime and the compiler interface CommaRT to support the raising
of arbitrary exceptions.  In particular, support the system exceptions
PROGRAM_ERROR and CONSTRAINT_ERROR using this new facility.

The only visible change is that the default exception handler in the main
routine reports the name of the caught exception.