Support the raise statement.
This patch implements the parsing, type checking, and codegen aspects of raise
statements.  Since there are no user-defined exceptions yet, the only thing that
can be done is raise the predefined exceptions "Program_Error" and
"Constraint_Error".

There are no handlers yet.  All exceptions raised are caught by the landing pad
of the main routine.  In fact, this is the shortest hello world program that can
be written ATM:

   domain D with
      procedure P;
   add
      procedure P is
      begin
         raise Program_Error with "hello, world!";
      end P;
   end D;

One step at a time...