Compile the runtime with -fexceptions.
Also, call _Unwind_RaiseException instead of _Unwind_Resume when re-throwing an
exception.  Ideally we should be calling _Unwind_Resume directly from the
generated code.  However, the handlers are given a pointer to a Comma exception
object instead of a pointer to the systems _Unwind_Exception object, and calling
into the runtime is easier than doing the necessary pointer tricks directly.