Improve the test suite.
The previous approach to testing used a very simple mechanism.  The exit status
of the driver and "MUST-FAIL" flags embedded in test comments were used to
determine if a test passed or failed.  This worked but was too "coarse".

Instead of MUST-FAIL flags, the testsuite now checks for EXPECTED-ERROR and
EXPECTED-WARNING tags.  These tags are embedded in comments and provide regular
expressions to match against the compiler diagnostics.  The location of the tags
is used to infer which diagnostic they intend to match.  There are additional
notes in testsuite/lib/comma-dg.exp.

This is certainly an improvement, but is not ideal.  The compiler will
eventually support this kind of testing in time making the TCL scripts obsolete.