Codegen architecture enhancements.
This patch, in essence, renames CodeGenCapsule to CGContext.  The primary
motivation for this change is to present the code generators "top level state"
thru a common interface.  This will allow domains, packages and subroutines to
establish the context under which the code generator functions.  Although this
patch does not introduce any significant change in functionality, it does
represent an important conceptual step.