Large patch extending record type checking and codegen.
The main contribution of this patch is support for simple array expressions.
Test cases are pending which illustrate the current capabilities.

In addition, simple object renaming declarations are supported as well.  The
main use of these declarations ATM is to effectively alias inj's and prj's
within the bodys of domains.  For example, given a carrier type named Rep we can
write:

   procedure P (X : in out %) is
     A : Rep renames inj(X);
     ...
   end P;

with the effect that the name A is equivalent semanticly to the expression
inj(X).  Note however that renaming declarations are not macros since the
elaboration of such a declaration includes the evaluation the associated
expression.