This patch adds initial support for incomplete type declarations.
An incomplete type declaration is of the form "type Foo;".  The main motivation
for incomplete types is that they will be used for forming self-recursive data
structures once access types are implemented.  In other words, this is the first
patch in a series that is moving towards support for things like the following:

   type Cell;
   type Cell_Access is access Cell;

   type Cell is record
      Elem : Integer;
      Next : Cell_Access;
   end record;