Only rewrite an instances representation instead of the entire private body.
The previous approach taken had obvious inefficiencies.  The full contents of an
AddDecl were being rewritten to ensure that the carrier declaration was fully
resolved.  This patch expands DeclRewriter to recursively and selectively
rewrite any dependent declarations a type may have, thus allowing a domain
instance to compute it's own version of the carrier (and any dependent types)
and nothing more.